[level 0] Title: 옷가게 할인 받기, Time: 0.02 ms, Memory: 73.5 MB -BaekjoonHub

+++ 프로그래머스/lv0/120818. 옷가게 할인 받기/옷가게 할인 받기.java
@@ -3,6 +3,6 @@ class Solution {
         if(price >= 500000) return (int)(price - (price*0.2));
         else if(price >= 300000) return (int)(price - (price*0.1));
         else if(price >= 100000) return (int)(price - (price*0.05));
-        else return (int)price;
+        else return price;
     }
 }
